requirements.txt

Flask
python-dotenv
requests


https://app.noest-dz.com/